Add semantic segmentors as encoders.
Treat semseg maps as a feature map by encoding them into onehot encoding and aggregating such that every dimension represents the sum of votes for a particular class. This makes it so we do not need to change any mapper to accomodate semsegs. It also allows us to project ground truth segmentations onto rays and voxels to improve the mapping independent of encoder.